requirements.txt
discord.py==2.2.2
openai==0.27.0
python-dotenv==0.19.2
git+https://github.com/ytdl-org/youtube-dl.git@master#egg=youtube_dl
PyNaCl==1.5.0
mysql-connector-python==8.0.32
APScheduler==3.10.1
python-dateutil==2.8.2